requirements.txt
aiohttp 
aiosignal 
asgiref==3.5.0
async-timeout 
asynctest==0.13.0
attrs 
bleach==4.1.0
blinker==1.4
brotlipy==0.7.0
cachetools==5.0.0
certifi==2021.10.8
cffi 
charset-normalizer==2.0.12
click==8.1.2
colorama 
cryptography==36.0.2
defusedxml==0.7.1
Django==3.2.12
django-allauth==0.50.0
django-ckeditor==6.3.0
django-crm==0.9.0
django-js-asset==2.0.0
django-summernote==0.8.20.0
django-tinymce==3.4.0
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
frozenlist 
google-auth==2.6.2
google-auth-oauthlib==0.5.1
greenlet==1.1.2
idna 
importlib-metadata 
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
multidict 
oauth==1.0.1
oauthlib 
packaging==21.3
Pillow==9.1.0
pyasn1 
pyasn1-modules==0.2.8
pycparser 
PyJWT==2.3.0
pyOpenSSL
pyparsing==3.0.7
PySocks 
python3-openid==3.2.0
pytz==2022.1
requests 
requests-oauthlib==1.3.1
rsa==4.8
six 
SQLAlchemy==1.4.34
sqlparse==0.4.2
typing_extensions 
urllib3==1.26.9
webencodings==0.5.1
Werkzeug==2.1.1
win-inet-pton 
wincertstore==0.2
yarl 
zipp==3.8.0
